+++ pkg/apis/piraeus/v1alpha1/linstor_types.go
@@ -17,7 +17,9 @@ limitations under the License.
 
 package v1alpha1
 
-import lapi "github.com/LINBIT/golinstor/client"
+import (
+	lapi "github.com/LINBIT/golinstor/client"
+)
 import lapiconst "github.com/LINBIT/golinstor"
 
 // NodeStatus simple status of the node in the linstor cluster.
@@ -137,4 +139,14 @@ func (lsc *LinstorSSLConfig) Type() string {
 	} else {
 		return lapiconst.ValNetcomTypeSsl
 	}
+}
+
+type LinstorClientConfig struct {
+	// Name of the secret containing:
+	// (a) `ca.pem`: root certificate used to validate HTTPS connections with Linstor (PEM format, without password)
+	// (b) `client.key`: client key used by the linstor client (PEM format, without password)
+	// (c) `client.cert`: client certificate matching the client key (PEM format, without password)
+	// If set, HTTPS is used for connecting and authenticating with linstor
+	// +kubebuilder:validation:Optional
+	LinstorHttpsClientSecret string `json:"linstorHttpsClientSecret"`
 }